scarper/requirements.txt
scrapy
pymongo
boto3
requests
python-dotenv
